HW4/pyproject.toml
[project]
name = "slam-hw-icp"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "numpy==1.26.4",
    "matplotlib",
    "Pillow",
    "rerun-sdk",
    "tyro",
    "tqdm",
    "numpy-quaternion",
    "pypng>=0.20220715.0",
]
